// src/app/Enum/heroes.enum.ts
export enum HeroesEnum {
    Iron_Man = 1,
    Thor = 2,
    Hulk = 3,
    Vision = 4,
    Black_Widow = 5,
    War_Machine = 6,
    Falcon = 7,
    Black_Panter = 8,
    Ant_Man = 9,
    Spider_Man = 10,
    Winter_Soldier = 11,
    Capitan_America = 12,
    Wanda_Maximoff = 13,
    Doctor_Strange = 14,
    Capitan_Marvel = 15,
    Wolverine = 16,
    La_Mole = 17,
    La_Antorcha_Humana = 18,
    La_Mujer_Invisible = 19,
    El_Hombre_Elastico = 20,
    Ciclope = 21,
    Tormenta = 22,
    Jean_Grey = 23,
    Profesor_X = 24
}